requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.118.0
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
openai==1.12.0
schedule==1.2.1
mysql-connector-python==8.3.0
SQLAlchemy==2.0.27
alembic==1.13.1
pydantic==2.6.1
python-dateutil==2.8.2
requests==2.31.0 